move components into src
