+++ src/pages/events/[id].tsx
@@ -3,7 +3,7 @@ import Link from "next/link"
 import { useRouter } from "next/router"
 import useSwr from "swr"
 
-import Layout from "../../../components/Layout"
+import Layout from "../../components/Layout"
 import { EventDocument } from "../../../interfaces"
 
 const EventById: NextPage = () => {
