+++ src/pages/events/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 import Link from "next/link"
-import Layout from "../../../components/Layout"
+import Layout from "../../components/Layout"
 import useSwr from "swr"
 
 import { EventDocument } from "../../../interfaces"
